Adds the Metropolis-Hastings algorithm
This Metropolis-Hastings version still is the slowest version, since it
must compute the determinant every time. Still faster than the algorithm
described by Kulesza et al.

In the next steps, I'll change the MH algorithm to compute the
probabilities using the inverse, which is updated iteratively so it's
not computationally expensive.